Newgrounds songs, moderator actions, list commands, votes actions

+++ incl/levels/deleteGJLevelUser.php
@@ -1,5 +1,4 @@
 <?php
-require __DIR__."/../../config/misc.php";
 require_once __DIR__."/../lib/mainLib.php";
 require_once __DIR__."/../lib/security.php";
 require_once __DIR__."/../lib/exploitPatch.php";
@@ -8,8 +7,12 @@ $sec = new Security();
 
 $person = $sec->loginPlayer();
 if(!$person["success"]) exit(CommonError::InvalidRequest);
+$userID = $person['userID'];
 
 $levelID = Escape::number($_POST['levelID']);
+
+$level = Library::getListByID($levelID);
+if(!$level || ($level['userID'] != $userID && !Library::checkPermission($person, 'commandDelete'))) exit(CommonError::InvalidRequest);
 
 Library::deleteLevel($levelID, $person);
